requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
playwright>=1.49.0
pyyaml>=6.0
jinja2>=3.1.0
python-multipart>=0.0.12
